% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlpanel.R
\name{controlpanel}
\alias{controlpanel}
\title{controlpanel}
\usage{
controlpanel(opts, factorList = c(), multiSelect = FALSE, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{opts}{a list of options to give the control panel}

\item{factorList}{a character vector of factors that will be made available in the Add Factor dropdown menu}

\item{multiSelect}{TRUE/FALSE if you want the user to be able to select more than one panel at a time}
}
\value{
a controlpanel html widget
}
\description{
Creates a control panel for adjusting factors and weights
}
