% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlpanel.R
\name{renderControlpanel}
\alias{renderControlpanel}
\title{Shiny render for controlpanelOutput}
\usage{
renderControlpanel(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a controlpanel}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.}
}
\description{
Shiny render for controlpanelOutput
}
